discovery: set GossipTimestampRange more reliably
In this commit the GossipTimestampRange that is sent to the syncer peer
is now set to have a "FirstTimestamp" based on the last update received.
Before, it was set to ask only updates that were created during the last
hour,  making it highly possible for a peer to miss updates after being
offline for a while.

With this new commit we use the last update time, which will reduce the
chance for missing updates, enableing a more reliable sync.